# Vectorize `search_n` for CPUs with SSE4.2 but not AVX2 support; handle AVX2 tail